requirements.txt
speechrecognition
pyttsx3
langchain
langchain-openai
openai
python-dotenv
